backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
reportlab==4.0.4
openai==1.3.0
anthropic==0.7.0